todo app functionality
had no idea where to go, but seems like i need to work on these projects a bit more to explore various data types and their use cases.
in this i used simple array and used push method to add the value to the of array and then cleaned the default todo value.